Fixed node_modules docker
Added anonymous volume to /app/node_modules to the development Compose file. This contains a copy of /app/node_modules and mounts it within the container, overriding the file system from the bind mount from the host which is hiding the node_modules directory installed in the image.
Created named volume

replace node_modules anonymous volume with a named one