Adding NgRx 'Redux' style state management
The global Store effectively replaces the need to store or maintain any state through the use of services. Services may still be used for centralising logic.  The Store is injected into pur components which select from and dispatch to the Store, which makes for much leaner code than relying on the Subject-Service patern we used previously.