Convert PlaceInput to a stateful component
Here we see some benefits of using a stateful component over a stateless one. The two-way data binding and checks we do on the input can be self contained within the component. All we need to do is output the value to a prop which can be bound to by the parent.