Add stacked navigation and side drawer
We added stacked navigation to FindPlace, to go from a place in PlaceList to a PlaceDetail view (which we now updated to a screen). We get a back button to navigate back in the stack out-of-the-box on iOS and Android. We next connected PlaceDetails to the Redux store to allow us to dispatch deletePlace action directly from the component. Doing so we pop the navigation stack to go back to the FindPlace screen.

After this we added a side drawer to the MainTabs screen. Note it is super buggy and had to put a work around in to fix the width in config, but it should be possible to set the drawer width dynamically to adjust on rotation.