Work on input elements in Auth screens
We have configured the email input to use the email-address keyboardType and not autoCapitalize or authCorrect. The password inputs use secure mode to hide the password.

We  configured the keyboard to pad the screen contents to make the inputs visable and also be closed by clicking any where on the screen