Install firebase-tools and create cloud Function
The important part is in functions/index.js. In here we create our cloud function which gets deployed to Firebase Functions compute server. We can target this function via REST api, to send our image data to, as the Firebase storage does not have a web API of its own and the Firebase SDK is currently broken in React Native.