Retrieve data with getPlaces thunk
The getPlaces thunk fetches data from Firebase and also does some useful transformations of that data before it is stored within the app. To store the data we have created a SET_PLACES action which we dispatch from the getPlaces thunk. 

As we are now managing the state on the Firebase server, there is no need to have the ADD_PLACE action, so we have removed this from the app.