Create authGetToken helper action
The authGetToken helper action makes the code more DRY, but as it returns a promise it provides some nice syntax in our other actions that require the token. In those actions we can chain promise operators, then and catch, from the action.

For now, the addPlace action will not function correctly as it is using the Firebase Functions API, so our token will not work.