Fill sidenav with navigation links
The sidenav is used on extra small devices. We use a mat-nav-list to style the list of items. The last item is a button, so we wrap this in a mat-list-item element. We add a style to the nav item text to ensure it is spaced nicely. We also add click listeners to close the sidenav when an option is clicked.