Adding header element with navbar
Here we have made use of the display property 'inline-block' value to make the elements in the header flow horizontally: the brand and the list of links.

To get the nav to fill the remaining width we had to do a hacky method using 'calc' to calculate the full width minus the width of the brand div. We also saw that whitespace in the HTML code between the div and nav actually contributed to the width! To work around this issue, we subtracted a little extra width off the nav.