Fix hover border style
The higher specificity of the rule used to style #free element caused the left-border to override the hover style. We could have used !important on the original style to ensure the hover effect was fixed. However, this breaks the concept on specifity and should be used rarely. Instead, we reapplied our styling to the element in a clean, though slightly non-DRY way.