Add brand logo using <img> and image styling
Compared to styling background images, there is not much more you can do to style a normal <img> image, apart from settings its size and adding space using margin/padding.

In addition, to size the image relative to the container, the container must be a inline-block or block level element.