Add grayscale filter to package background image
We've improved the background image on the package page by sizing it using 'cover' so it fills the container, and positioning it centrally using the 'center' positioning value.

In addition, we have also applied a grayscale filter to give the image a gray hint, making it blend into the background nicely.