Add a backdrop element
The backdrop element will be used in conjunction with the modal in the next section.

We explored the unexpected effects of using relative positioning with percentage height. The background did not determine a height (it was zero) because the body element did not explicitly define its height through CSS. This could be fixed by adding height: 100% to html and body elements.

We also looked at using absolute positioning on the backdrop. The html, body height was no longer required. This was because the containing block for absolute is the ancestor which is non-static which is neither body/html. In this case, the behaviour falls back to fixed positioning, i.e, it uses the viewport.

Lastly, we finished the backdrop using fixed positioning. This ensure the background sizes were calculated relative to the viewport and it stays in place when scrolling. In addition, we set top and left to zero to overcome margin collapsing issues.