Add max-width to image container
This is the beginnings of responsive design, i.e. we are controlling how the image behaves across a range of screen sizes. Adding an absolute max-width in conjunction with a percentage width, allows the image to scale freely up to a maximum width, at which point it would look too large if it got any larger.