Improve the Customer Page
We have follows the standard design again, on smaller devices layout vertically with the text positioned below. Therefore, our standard CSS we don't need as much code for margins and width.

We've created a smooth transition between our breakpoint, by setting the customer image max-width to 40rem. Therefore, it'll stop stretching once we go into the larger screens.

We've also set a max-width on the testimonial row and added auto margin to the left and right. Therefore, when it reaches its max width it will become centred horizontally.