Improve Packages Page
We have made the packages page a bit more responsive. We've set the max width of the package to 1000px once in the larger screen breakpoint. This prevents the package from stretching too wide. In addition, we've used horizontal auto margins to centre the packages once this kicks in.

We also had to fix some issues with the borders on the packages, which were missing on the left and right side on our desktop design.