Clean up Footer and normalise pages
Here we have normalised our pages, since each page has a navbar it makes sense to target the main element to add margin-top, to stop the navbar overlapping the main area, rather than let each page handle that itself.

In addition, we calculate the min-height of the main area so that the footer is always pushed to the bottom of the page. This works since the main area cannot shrink less than the viewport size minus the size of the navbar and footer.