Fix issue caused by html/body height
The issue arises due to margin collapsing with the main element. The main element has a top margin to stop it from extending underneath the header. However, this margin collapses with the body element, so the body element's content box (which is 100% height) gets pushed down. This results in vertical scroll bars and space below the footer.

I've removed the height 100% on html/body but do not fully understand why this issue should arise in the first place