Added form validation feedback
We've created a class we can add to the form elements manually or dynamically via JavaScript which shows some visual feedback that the control is invalid.

We also use the CSS :invalid pseudo-class selector which is selects elements by there invalid state. This is a HTML property which is set by the browser for inputs such as email, url and required if they are invalid.