WIP: Add GBFS spider
Working the GBFS spider. I'm using attrs to strongly declare the structure of the GBFS JSON responses, which are parsed by cattrs. 

The main difficultly with bikeshare scraping is the fact there are multiple items to process.

In addition, it will be more practical and more efficient to batch process the items, i.e. rather than yield each item individually, either do a lot of processing up front in the spider or yield a list of items and process them all together in the pipeline.