Add Station model and fix issue with Alembic env
Added Station model with foreign key to System table.

I had to remove the model imports from models/__init__.py, which I thought could be used to 'register' the models for Alembic to easily find. This was because the imports inteferred with normal useage in SqlAlchemy as it would say 'Table ... is already defined for this MetaData...' when you try to import a model