Add Scraper SqlAlchemy model
The Scraper table contains the name of each scraper that is run. The System and Station model as a FK to the Scraper model. This allows every system and station that is scraped and saved to the DB to be associated with the scraper that created it. The source_id and scraper_id fields on the System and Station model have a joint unique constraint. This provides a way to uniquely identify an object using its source_id, which will be unique within the 'dataset' managed by the scraper.