WIP: Write items to the DB
I've decided I will persist each to the DB within the pipeline rather than in the spider code. This is because, there will only be one spider running per process (as I will be scheduling individual spiders to run from Airflow).

Therefore, I've started work on processing each type of item. The data yielded from the spider will contain a 'item_type' field to distinguish them. In addition, I've created a BaseItem which contains cattr (de)serialisation methods.

This allows the item to be re-inflated in the pipeline again. This is hugely ineffecient but can live with it for now until I figure out how to back the attr-class properly subclass Scrapy BaseItem