Insert or update System and add mypy
The pipeline now inserts or updates each SystemItem to the database. It first queries the DB for the System using its source_id. If no system if returned, it will insert a record.

I've also added mypy to project. The mypy plugin in PyCharm can be run to perform static type checking.