Persist stations to the database
The stations are now also persisted to the database using the update_or_create method. An extra step is required to link up the station to the persisted system.

I've made a change in the gbfs spider to not pass through the source_id for the SystemItem. This is because gbfs sources only contain a single system. The source_id's global uniqueness is 'left down to the publishers' anyway. In addition, the station feeds do not contain the system's source_id. So it would be difficult to get the source_id from the system data in the asynchronous environment.