Fix type errors and move dev pipenv packages
The generic factory function on the GbfsBaseModel (which returns type T), does not typecheck on constructed instances, as the types on the subclass are missing. I've used to documented 'bound' property on the TypeVar declaration. This means the inferred type must be a subclass of the boundary type.

In addition, I also learned about 'ClassVar' which is used to annotation class variables (static members). I'm not sure whether attr includes these in __init__, but it would be good if it didn't.