Used skew and scale to improve the customers page
We used skw and scale transformations to improve the look of the customer images on the customer page. We have used skew to give the images are slatted diamond shape. We applied a skew transformation to the image itself to negate the distortion and increased the scale of the image to fill the container.