Applying transformation to containers and 'preserving 3D-ness'
When we apply a rotation to a container, we see that only a flatten 2D image of the container's elements is transformed. When rotating the container to be edge on, everything disappears. This is because the 3D space within the container is 'flattened' onto the surface of the container. We can preserve this 3D space using 'transform-style: preserve-3d'.