Transition on backdrop with 'display: none' workaround
In order to give the backdrop a transition effect, we must work around issues with display: none. Transitions do not work if they occur at the same time as display is changed. Therefore, in our JavaScript we add a setTimeout callback to change the appropriate property, i.e. when opening the backdrop, we add the open class after a small delay while when closing, we remove the open class and then set display: none after transition has finished