figured out the issue
It causes problems when you import a type from a module to use client-side
if that module contains other dependencies that cannot be bundled for the
client. This occurs when trying to call the getOrderStatus function in the
client-side SDK because that module imports seemingly harmless types, such
as the OrderStatus enum. This enum is defined in the Temporal package which
clearly causes problems for webpack.