fix: manager EKS API
I found that I could not log into the EKS cluster when assuming the eks-admin
IAM Role, as I got "unauthorised" errors. The issue was I had added the manager
IAM User to the `my-admin` k8s group via the aws_eks_access_entry resource
rather than the `eks-admin` IAM Role via the API.

The manager IAM User does not have an IAM policy like the developer user has
that allows them to DescribeCluster (or do anything with EKS). So even though
the manager IAM User was added to the k8s group, they wasn't allowed to access
the EKS cluster due to IAM permissions. The manager IAM User instead is allowed
to assume the eks-admin IAM Role, which has the IAM Policy to access the cluster.
By updating the EKS API to add the eks-admin IAM Role to the my-admin cluster group,
I was able to both authenticate against the EKS cluster and was authorised to get
nodes, etc.